1.2dev: merged [14250-14251] from 1.0-stable (fix for #12118)
git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@14252 af82e41b-90c4-0310-8c96-b1721e28e2e2

+++ trac/versioncontrol/api.py
@@ -234,6 +234,13 @@ class DbRepositoryProvider(Component):
         if is_default(target):
             target = ''
         rm = RepositoryManager(self.env)
+        repositories = rm.get_all_repositories()
+        if target not in repositories:
+            raise TracError(_("Repository \"%(repo)s\" doesn't exist",
+                              repo=target or '(default)'))
+        if 'alias' in repositories[target]:
+            raise TracError(_('Cannot create an alias to the alias "%(repo)s"',
+                              repo=target or '(default)'))
         with self.env.db_transaction as db:
             id = rm.get_repository_id(reponame)
             db.executemany(
@@ -247,6 +254,11 @@ class DbRepositoryProvider(Component):
         if is_default(reponame):
             reponame = ''
         rm = RepositoryManager(self.env)
+        repositories = rm.get_all_repositories()
+        if any(reponame == repos.get('alias')
+               for repos in repositories.itervalues()):
+            raise TracError(_('Cannot remove the repository "%(repos)s" used '
+                              'in aliases', repos=reponame or '(default)'))
         with self.env.db_transaction as db:
             id = rm.get_repository_id(reponame)
             db("DELETE FROM repository WHERE id=%s", (id,))
@@ -262,6 +274,13 @@ class DbRepositoryProvider(Component):
         if is_default(new_reponame):
             new_reponame = ''
         rm = RepositoryManager(self.env)
+        if reponame != new_reponame:
+            repositories = rm.get_all_repositories()
+            if any(reponame == repos.get('alias')
+                   for repos in repositories.itervalues()):
+                raise TracError(_('Cannot rename the repository "%(repos)s" '
+                                  'used in aliases',
+                                  repos=reponame or '(default)'))
         with self.env.db_transaction as db:
             id = rm.get_repository_id(reponame)
             if reponame != new_reponame:
